Explanation
This is an example of a simple counting app using python. It has been coded using flet module which is a framework to build interactive applications. In other words, it is an easier way to create FLUTTER apps. Flet controls are based on Flutter by Google. Counter app is an open source tally counter used as applications in android, apple ,etc. . They have been upgraded into calorie measurers, tally counters and similar apps.

*Installing flet module (if not already installed) is the initial step. Use the common way ie. pip function.
*Import the module and its functions required to create the app.
*Describe the necessary methods to place the title as required.
*Initialize initial value as zero and define the add_one and minus_one functions.
*Commands to display the data in an app format is given.

Above instructions guide in creating a mini app using flet module. It is a really interesting project to take!!!!!!